+++ Deroute/Collab/CollabSettings.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Globalization;
 
 namespace DerouteSharp.Collab
 {
