+++ Deroute/Collab/EntityConverter.cs
@@ -1,9 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Drawing;
-using System.Linq;
-using System.Windows.Forms;
-using System.Runtime.InteropServices;
 
 namespace DerouteSharp.Collab
 {
