+++ Deroute/Collab/UI/CollabStatusPanel.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Drawing;
-using System.Linq;
 using System.Windows.Forms;
-using DerouteSharp.Collab;
 
 namespace DerouteSharp.Collab.UI
 {
